requirements.txt
paho-mqtt
gnocchiclient
jsonschema
